Refactored sidecar loading; Added more logging for sidecar loading; Added clear buttons in material editor;

+++ TombLib/TombLib.Forms/Forms/FormMaterialEditor.cs
@@ -8,6 +8,7 @@ using System.Numerics;
 using System.Windows.Forms;
 using TombLib.LevelData;
 using TombLib.Utils;
+using static TombLib.LevelData.Compilers.TombEngineTexInfoManager;
 
 namespace TombLib.Forms
 {
@@ -20,7 +21,8 @@ namespace TombLib.Forms
 
 		private readonly Color _correctColor;
 		private readonly Color _wrongColor;
-		private static readonly Logger _logger = LogManager.GetCurrentClassLogger();
+
+		private bool _saveXml = false;
 
 		public FormMaterialEditor(string texturePath)
 		{
@@ -41,6 +43,18 @@ namespace TombLib.Forms
 			tbAmbientOcclusionMapPath.Text = _materialData.AmbientOcclusionMap;
 			tbEmissiveMapPath.Text = _materialData.EmissiveMap;
 
+			if (!string.IsNullOrEmpty(_materialData.NormalMap))
+				tbNormalMapPath.BackColor = (_materialData.IsNormalMapFound ? _correctColor : _wrongColor);
+
+			if (!string.IsNullOrEmpty(_materialData.AmbientOcclusionMap))
+				tbAmbientOcclusionMapPath.BackColor = (_materialData.IsAmbientOcclusionMapFound ? _correctColor : _wrongColor);
+
+			if (!string.IsNullOrEmpty(_materialData.SpecularMap))
+				tbSpecularMapPath.BackColor = (_materialData.IsSpecularMapFound ? _correctColor : _wrongColor);
+
+			if (!string.IsNullOrEmpty(_materialData.EmissiveMap))
+				tbEmissiveMapPath.BackColor = (_materialData.IsEmissiveMapFound ? _correctColor : _wrongColor);
+
 			LoadTexturePreview(_materialData.ColorMap, picPreviewColorMap);
 			LoadTexturePreview(_materialData.NormalMap, picPreviewNormalMap);
 			LoadTexturePreview(_materialData.SpecularMap, picPreviewSpecularMap);
@@ -55,24 +69,33 @@ namespace TombLib.Forms
 
 		private void LoadTexturePreview(string path, PictureBox pictureBox)
 		{
-			try
+			if (string.IsNullOrEmpty(path))
 			{
-				if (!string.IsNullOrEmpty(path))
-				{
-
-					pictureBox.Image?.Dispose();
-					pictureBox.Image = ImageC.FromFile(path).ToBitmap();
-					pictureBox.BackgroundImage = TombLib.Properties.Resources.misc_TransparentBackground;
-					pictureBox.Tag = null;
-					pictureBox.BackColor = _correctColor;
-				}
-			}
-			catch (Exception exc)
-			{
+				pictureBox.Image?.Dispose();
 				pictureBox.Image = null;
-				pictureBox.BackgroundImage = null;
-				pictureBox.Tag = exc;
-				pictureBox.BackColor = _wrongColor;
+				pictureBox.BackgroundImage = TombLib.Properties.Resources.misc_TransparentBackground;
+			}
+			else
+			{
+				try
+				{
+					if (!string.IsNullOrEmpty(path))
+					{
+
+						pictureBox.Image?.Dispose();
+						pictureBox.Image = ImageC.FromFile(path).ToBitmap();
+						pictureBox.BackgroundImage = TombLib.Properties.Resources.misc_TransparentBackground;
+						pictureBox.Tag = null;
+						pictureBox.BackColor = _correctColor;
+					}
+				}
+				catch (Exception exc)
+				{
+					pictureBox.Image = null;
+					pictureBox.BackgroundImage = null;
+					pictureBox.Tag = exc;
+					pictureBox.BackColor = _wrongColor;
+				}
 			}
 		}
 
@@ -98,45 +121,48 @@ namespace TombLib.Forms
 
 		private void butOK_Click(object sender, EventArgs e)
 		{
-			string externalMaterialDataPath = Path.Combine(
-					 Path.GetDirectoryName(_texturePath),
-					 Path.GetFileNameWithoutExtension(_texturePath) + ".xml");
-
-			var materialData = new MaterialData();
-
-			materialData.Type = (MaterialType)comboMaterialType.SelectedIndex;
-
-			materialData.ColorMap = _texturePath;
-			materialData.NormalMap = tbNormalMapPath.Text;
-			materialData.SpecularMap = tbSpecularMapPath.Text;
-			materialData.EmissiveMap = tbEmissiveMapPath.Text;
-			materialData.AmbientOcclusionMap = tbAmbientOcclusionMapPath.Text;
-
-			switch (materialData.Type)
+			if (_saveXml)
 			{
-				case MaterialType.Opaque:
-					materialData.Parameters0 = new Vector4(
-							(float)nmNormalMapStrength.Value,
-							(float)nmSpecularIntensity.Value,
-							0.0f,
-							0.0f);
-					break;
+				string externalMaterialDataPath = Path.Combine(
+						 Path.GetDirectoryName(_texturePath),
+						 Path.GetFileNameWithoutExtension(_texturePath) + ".xml");
+
+				var materialData = new MaterialData();
+
+				materialData.Type = (MaterialType)comboMaterialType.SelectedIndex;
+
+				materialData.ColorMap = _texturePath;
+				materialData.NormalMap = tbNormalMapPath.Text;
+				materialData.SpecularMap = tbSpecularMapPath.Text;
+				materialData.EmissiveMap = tbEmissiveMapPath.Text;
+				materialData.AmbientOcclusionMap = tbAmbientOcclusionMapPath.Text;
+
+				switch (materialData.Type)
+				{
+					case MaterialType.Opaque:
+						materialData.Parameters0 = new Vector4(
+								(float)nmNormalMapStrength.Value,
+								(float)nmSpecularIntensity.Value,
+								0.0f,
+								0.0f);
+						break;
+				}
+
+				try
+				{
+					File.Delete(externalMaterialDataPath);
+					MaterialData.SaveToXml(externalMaterialDataPath, materialData);
+					MaterialFileName = externalMaterialDataPath;
+				}
+				catch (Exception)
+				{
+					DarkMessageBox.Show(this, $"An error occurred while saving XML material file to '{externalMaterialDataPath}'", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+					return;
+				}
 			}
 
-			try
-			{
-				File.Delete(externalMaterialDataPath);
-				MaterialData.SaveToXml(externalMaterialDataPath, materialData);
-				MaterialFileName = externalMaterialDataPath;
-
-				DialogResult = DialogResult.OK;
-				Close();
-			}
-			catch (Exception ex)
-			{
-				DarkMessageBox.Show(this, $"An error occurred while saving XML material file to '{externalMaterialDataPath}'", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-				return;
-			}
+			DialogResult = DialogResult.OK;
+			Close();
 		}
 
 		private void comboMaterialType_SelectedIndexChanged(object sender, EventArgs e)
@@ -150,7 +176,9 @@ namespace TombLib.Forms
 			if (!string.IsNullOrEmpty(texturePath))
 			{
 				tbNormalMapPath.Text = texturePath;
+				tbNormalMapPath.BackColor = _correctColor;
 				LoadTexturePreview(texturePath, picPreviewNormalMap);
+				_saveXml = true;
 			}
 		}
 
@@ -160,7 +188,9 @@ namespace TombLib.Forms
 			if (!string.IsNullOrEmpty(texturePath))
 			{
 				tbAmbientOcclusionMapPath.Text = texturePath;
+				tbAmbientOcclusionMapPath.BackColor = _correctColor;
 				LoadTexturePreview(texturePath, picPreviewAmbientOcclusionMap);
+				_saveXml = true;
 			}
 		}
 
@@ -170,7 +200,9 @@ namespace TombLib.Forms
 			if (!string.IsNullOrEmpty(texturePath))
 			{
 				tbEmissiveMapPath.Text = texturePath;
+				tbEmissiveMapPath.BackColor = _correctColor;
 				LoadTexturePreview(texturePath, picPreviewEmissiveMap);
+				_saveXml = true;
 			}
 		}
 
@@ -180,8 +212,64 @@ namespace TombLib.Forms
 			if (!string.IsNullOrEmpty(texturePath))
 			{
 				tbSpecularMapPath.Text = texturePath;
+				tbSpecularMapPath.BackColor = _correctColor;
 				LoadTexturePreview(texturePath, picPreviewSpecularMap);
+				_saveXml = true;
 			}
+		}
+
+		private void butClearNormalMap_Click(object sender, EventArgs e)
+		{
+			if (DarkMessageBox.Show(this, "Do you really want to clear the normal map?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+			{
+				tbNormalMapPath.Text = "";
+				tbSpecularMapPath.BackColor = this.BackColor;
+				LoadTexturePreview("", picPreviewNormalMap);
+				_saveXml = true;
+			}
+		}
+
+		private void butClearAmbientOcclusionMap_Click(object sender, EventArgs e)
+		{
+			if (DarkMessageBox.Show(this, "Do you really want to clear the ambient occlusion map?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+			{
+				tbAmbientOcclusionMapPath.Text = "";
+				tbSpecularMapPath.BackColor = this.BackColor;
+				LoadTexturePreview("", picPreviewAmbientOcclusionMap);
+				_saveXml = true;
+			}
+		}
+
+		private void butClearEmissiveMap_Click(object sender, EventArgs e)
+		{
+			if (DarkMessageBox.Show(this, "Do you really want to clear the emissive map?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+			{
+				tbEmissiveMapPath.Text = "";
+				tbSpecularMapPath.BackColor = this.BackColor;
+				LoadTexturePreview("", picPreviewEmissiveMap);
+				_saveXml = true;
+			}
+		}
+
+		private void butClearSpecularMap_Click(object sender, EventArgs e)
+		{
+			if (DarkMessageBox.Show(this, "Do you really want to clear normal map?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+			{
+				tbSpecularMapPath.Text = "";
+				tbSpecularMapPath.BackColor = this.BackColor;
+				LoadTexturePreview("", picPreviewSpecularMap);
+				_saveXml = true;
+			}
+		}
+
+		private void nmNormalMapStrength_ValueChanged(object sender, EventArgs e)
+		{
+			_saveXml = true;
+		}
+
+		private void nmSpecularIntensity_ValueChanged(object sender, EventArgs e)
+		{
+			_saveXml = true;
 		}
 	}
 }
